+++ themoviedb/Services/TMDBService.swift
@@ -8,44 +8,89 @@
 import Foundation
 import Combine
 
+/// Protocols for the TMDB service
 protocol TMDBServiceProtocol {
     func fetchNowPlayingMovies(page: Int) async throws -> MovieList
+    func fetchMovieDetail(id: Int) async throws -> MovieDetail
+    func searchMovies(query: String) async throws -> MovieList
 }
 
+/// Final class to interact with the movie database API
 final class TMDBService: TMDBServiceProtocol {
+    
+    /// Singleton instance for global share
     static let shared = TMDBService()
-    private init() {}
-
-    private let baseURL = "https://api.themoviedb.org/3/"
-    private let apiKey = "4c5806a5de6fbed9b28f238484fede91"
-
-    // Endpoint to get the list of now playing movies
-    func fetchNowPlayingMovies(page: Int) async throws -> MovieList {
-        let urlStr = "\(baseURL)movie/now_playing?api_key=\(apiKey)&page=\(page)"
-        guard let url = URL(string: urlStr) else {
-            throw NetworkError.invalidURL
-        }
-        return try await NetworkManager.shared.request(url: url)
+    
+    private let networkManager: NetworkManager
+    
+    /// Allow injection for unit tests
+    init(networkManager: NetworkManager = .shared) {
+        self.networkManager = networkManager
     }
     
     
-    // Endpoint to get the movie details by movie ID
-    func fetchMovieDetail(id: Int) async throws -> MovieDetail {
-            let urlStr = "\(baseURL)movie/\(id)?api_key=\(apiKey)"
-        print("urlstr", urlStr)
-            guard let url = URL(string: urlStr) else {
-                throw NetworkError.invalidURL
-            }
-            return try await NetworkManager.shared.request(url: url)
+    /// Endpoint to get the list of now playing movies
+    /// Page for pagination implementation
+    func fetchNowPlayingMovies(page: Int) async throws -> MovieList {
+        let urlStr = "\(AppConstants.API.baseURL)movie/now_playing?api_key=\(AppConstants.API.apiKey)&page=\(page)"
+        
+        guard let url = URL(string: urlStr) else {
+            throw NetworkError.invalidURL
         }
+        
+        do {
+            return try await networkManager.request(url: url)
+        } catch {
+            throw mapError(error)
+        }
+    }
     
-    // Endpoint to get the movie details by movie ID
-    func searchMovies(query: String) async throws -> MovieList {
-            let urlStr = "\(baseURL)search/movie?api_key=\(apiKey)&query=\(query)"
-        print("searchURL", urlStr);
-            guard let url = URL(string: urlStr) else {
-                throw NetworkError.invalidURL
-            }
-            return try await NetworkManager.shared.request(url: url)
+    
+    /// Endpoint to get the movie details by movie ID
+    func fetchMovieDetail(id: Int) async throws -> MovieDetail {
+        let urlStr = "\(AppConstants.API.baseURL)movie/\(id)?api_key=\(AppConstants.API.apiKey)"
+        
+        guard let url = URL(string: urlStr) else {
+            throw NetworkError.invalidURL
         }
+        do {
+            return try await networkManager.request(url: url)
+        } catch {
+            throw mapError(error)
+        }
+        
+    }
+    
+    /// Endpoint to get the movie details by movie query for search
+    func searchMovies(query: String) async throws -> MovieList {
+        guard let encodedQuery = query.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) else {
+            throw NetworkError.invalidQuery
+        }
+        
+        let url = "\(AppConstants.API.baseURL)search/movie?api_key=\(AppConstants.API.apiKey)&query=\(encodedQuery)"
+        
+        guard let url = URL(string: url) else {
+            throw NetworkError.invalidURL
+        }
+        
+        do {
+            return try await networkManager.request(url: url)
+        } catch {
+            throw mapError(error)
+        }
+        
+    }
+    
+    /// mapError used for error handling
+    private func mapError(_ error: Error) -> NetworkError {
+        if let networkError = error as? NetworkError {
+            return networkError
+        }
+        return .unknown(error)
+    }
 }
+
+
+
+
+