+++ themoviedb/Utilities/APIKeys.swift
@@ -1,0 +1,11 @@
+//
+//  APIKeys.swift
+//  TheMovieDB
+//
+//  Created by Siva kumar Aketi on 15/05/25.
+//
+
+// APIKeys.swift - NOT committed to git but for now i am adding into my git
+struct APIKeys {
+    static let tmdbApiKey = "4c5806a5de6fbed9b28f238484fede91"
+}