+++ themoviedb/Utilities/Extensions.swift
@@ -1,7 +1,0 @@
-//
-//  Extensions.swift
-//  TheMovieDB
-//
-//  Created by Siva kumar Aketi on 12/05/25.
-//
-